requirements.txt
# Core dependencies for Email Processing CLI
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Python 3.11+ required
